moved products into its own folder
