+++ server/app/routes/index.js
@@ -3,7 +3,7 @@ var router = require('express').Router();
 module.exports = router;
 
 router.use('/members', require('./members'));
-router.use('/products', require('./products.js'));
+router.use('/products', require('./products'));
 
 // Make sure this is after all of
 // the registered routes!
