+++ server/app/routes/products/index.js
